refactor(scopus): move api interactions to dedicated module
